WIP: conversion to pure chat-bot
